+++ browser-extension/src/content.js
@@ -11,6 +11,7 @@
         IMGClose: 'shared/icons/close.svg',
         IMGPlus: 'shared/icons/plus.svg',
         IMGPreorder: 'shared/icons/preorder.svg',
+        IMGPlayAnywhere: 'shared/icons/play-anywhere.svg',
         IMGExpand: 'shared/icons/expand.svg',
         IMGCollapse: 'shared/icons/collapse.svg',
         CSSFilter: null // already declared in manifest.json's content_scripts.css